+++ aws-sdk-core/lib/aws-sdk-core.rb
@@ -68,6 +68,7 @@ module Aws
     LambdaPreview
     MachineLearning
     MarketplaceCommerceAnalytics
+    MarketplaceMetering
     OpsWorks
     RDS
     Redshift